Update templates to support license autoload.
- If autload is enabled, we disable the license job and mount the secret
  with the license as a volume mount on the consul server pods, client
pods and snapshot agent pods. Additionally we set the env CONSUL_LICENSE_PATH on these pods.

- When manageSystemACLs is enabled, we mount the secret only on the
  server pods as the consul clients and snapshot agents pods can grab
the license from the server during startup. This is done in order to
ensure only a server restart is required when the license file is
updates and not clients and snapshot agent as well.